Refactor endpoint trace graph finalize helper
